chore(get_service_logs): fold schema test into service_logs_test.go; drop CONTEXT.md (ENG-1317)
